+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
 - expose live sync rate limit settings in advanced options
 - fix potential error on comparing messages in live sync
 - fix escaping special characters for regex in exported HTML search functions
+- disable automatic update check and download popup in snapped version
 
 
 5.5, 2023-07-18